changed input to float

+++ 04.03/circumference.c
@@ -1,5 +1,5 @@
 #include "circle.h"
 
-float circumference(int radius){
+float circumference(float radius){
     return PI * radius * 2;
 }
